refactor: api 네이밍 1차 변경 (#99)
